Replace units string array with Unit enum
Using an enum for units is better than an array of strings since it
creates a tighter typing, but there are issues with this implementation.
Accessing values of the enum dynamically using key-value coding [1] means
we have to force downcast our value back to a UnitLength. If another dev
added a nonexistent case to our enum (i.e. a length value that did not
exist on the UnitLength class), our app will crash [2] when run. While
this isn't the greatest developer experience, it does at least mean the
developer will get a helpful error when they attempt to include a case
that doesn't exist on UnitLength. In general, casting is not an ideal
solution, and force casting (or downcasting in our case from type Any)
is an even worse solution. Perhaps a more ideal solution would somehow
conform our Unit enum to the UnitLength class properties (not even sure
this is possible) or conditionally run our casting (e.g. `if let`) to
help surface the error in our own way.

[1]: https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/KeyValueCoding/index.html
[2]: '[<NSUnitLength 0x7fff86d9aca0> valueForUndefinedKey:]: this class is not key value coding-compliant for the key foo.'